tests/wpt/tests/pointerevents/pointerevent_root_hit_test.html

<!doctype html>
<meta charset=utf-8>
<title></title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<style type="text/css">
    :root {
        pointer-events: none;
    }
</style>
<html>
    <body>
    </body>
</html>
<script>
    test(() => {
        let element = document.elementFromPoint(50, 50);
        assert_equals(element, document.documentElement);
    });
</script>
